+++ client/package.json
@@ -16,7 +16,9 @@
     "react-dom": "^16.13.1",
     "react-redux": "^8.1.1",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.1"
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.1",
+    "redux-thunk": "^2.4.2"
   },
   "scripts": {
     "start": "react-scripts start",
