--- client/src/reducer.js
@@ -1,33 +1,0 @@
-import {
-    UPDATE_PRODUCTS,
-    ADD_TO_CART,
-    UPDATE_CART_QUANTITY,
-    REMOVE_FROM_CART,
-    ADD_MULTIPLE_TO_CART,
-    UPDATE_CATEGORIES,
-    UPDATE_CURRENT_CATEGORY,
-    CLEAR_CART,
-    TOGGLE_CART,
-} from './utils/actions';
-
-const initialState = {
-    products: [],
-    cart: [],
-    categories: [],
-    currentCategory: "",
-    cartOpen: false,
-}
-
-
-// USE THE SYNTAX OF THIS CODE FROM REDUX DOCUMENTATION AS A GUIDE
-// Use the initialState as a default value
-export default function appReducer(state = initialState, action) {
-    // The reducer normally looks at the action type field to decide what happens
-    switch (action.type) {
-        // Do something here based on the different types of actions
-        default:
-        // If this reducer doesn't recognize the action type, or doesn't
-        // care about this specific action, return the existing state unchanged
-        return state
-    }
-}
