+++ client/src/utils/store.js
@@ -1,0 +1,4 @@
+import { configureStore } from '@reduxjs/toolkit'
+import appReducer from './reducers';
+
+export default configureStore(appReducer);
